heavy refactoring
