+++ src/components/graphs/contributions/contributionsGraphContainer.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Client from '../../api';
 import ChartWrapper from './chartWrapper';
 import { flatten } from 'lodash';
-import { Tabs, Tab } from 'react-bootstrap';
+import CampaignTabs from '../../tabs/campaignTabs'
 
 class ContributionsGraphContainer extends React.Component {
   constructor(props) {
@@ -10,12 +10,8 @@ class ContributionsGraphContainer extends React.Component {
     this.state = {
       chartData: [],
       successfulFetches: 0,
-      chartType: 'default',
-      activeTab: 'visualizer'
+      chartType: 'default'
     };
-    this.setActiveTab = this.setActiveTab.bind(this);
-    this.renderTabs = this.renderTabs.bind(this);
-    this.goFullScreen = this.goFullScreen.bind(this);
   }
 
   componentWillMount() {
@@ -47,129 +43,25 @@ class ContributionsGraphContainer extends React.Component {
     this.setState({ svg: svg });
   }
 
-  setActiveTab(tab) {
-    this.setState({ activeTab: tab });
-  }
+  render() {
+    let candidates = this.props.location.state
 
-  goFullScreen(tab) {
-    this.setState({ fullScreenTab: tab });
-  }
-
-  renderTabs(candidateCampaigns) {
-    const fullScreenStyle = {
-      top: 0,
-      bottom: 0,
-      position: 'fixed',
-      left: 0,
-      height: '100%',
-      width: '100%',
-      overflow: 'hidden',
-      zIndex: 99999
-    };
-
-    const { activeTab, fullScreenTab, chartData, chartType } = this.state;
-    const candidates = this.props.location.state;
+    let candidateCampaigns = candidates.map(candidate => {
+      for (let campaign of candidate.data.campaigns) {
+        return <p> {candidate.candidateName} - {campaign.raceType} </p>;
+      }
+    });
     const ids = flatten([
       ...candidates.map(can => {
         return can.data.campaigns.map(campaign => campaign.campaignId);
       })
     ]).join(':');
-    const vizTitle = (
-      <span>
-        Visualiations
-        {' '}
-        <button
-          style={{ padding: '0', paddingLeft: '2px', paddingRight: '2px' }}
-          onClick={() => this.goFullScreen('visualizer')}
-          disabled={activeTab !== 'visualizer'}
-          className="btn btn-info btn-xs"
-        >
-          <i className="fa fa-expand" />
-        </button>
-      </span>
-    );
-    const tableTitle = <span>Contribution Table</span>;
+    const { chartData, chartType } = this.state;
 
-    if (fullScreenTab === 'visualizer') {
+    // check for data integrity
+    if (chartData.length === candidates.length) {
       return (
-        <Tabs>
-          <Tab
-            id="keshif"
-            eventKey={1}
-            title={vizTitle}
-            onEntered={() => this.setActiveTab('visualizer')}
-          >
-            <div className="campaign-visualizer">
-              <iframe
-                style={fullScreenStyle}
-                src={
-                  'https://codefordc.org/campaign-finance-explorer/index.html?' +
-                    ids
-                }
-              />
-              <button
-                className="btn btn-info btn-xs"
-                onClick={() => this.goFullScreen('')}
-                style={{
-                  padding: '0',
-                  paddingLeft: '2px',
-                  paddingRight: '2px',
-                  position: 'fixed',
-                  top: '2px',
-                  right: '10px',
-                  zIndex: 999999
-                }}
-              >
-                <i className="fa fa-compress" />
-              </button>
-            </div>
-          </Tab>
-          <Tab
-            id="table"
-            eventKey={2}
-            title={tableTitle}
-            onEntered={() => this.setActiveTab('table')}
-          >
-            <div>
-              <h1> Contribution Comparison Chart </h1>
-              <h3> Selected Campaigns </h3>
-              {candidateCampaigns}
-              <ChartWrapper
-                chartInfo={chartData}
-                chartType={chartType}
-                onSvgCreate={this._setSvg.bind(this)}
-                candidates={candidates}
-              />
-            </div>
-          </Tab>
-        </Tabs>
-      );
-    }
-
-    return (
-      <Tabs>
-        <Tab
-          id="keshif"
-          eventKey={1}
-          title={vizTitle}
-          onEntered={() => this.setActiveTab('visualizer')}
-        >
-          <div className="campaign-visualizer">
-            <iframe
-              style={fullScreenTab === 'visualizer' ? fullScreenStyle : {}}
-              src={
-                'https://codefordc.org/campaign-finance-explorer/index.html?' +
-                  ids
-              }
-            />
-          </div>
-        </Tab>
-        <Tab
-          id="table"
-          eventKey={2}
-          title={tableTitle}
-          onEntered={() => this.setActiveTab('table')}
-        >
+        <CampaignTabs id = {ids}>
           <div>
             <h1> Contribution Comparison Chart </h1>
             <h3> Selected Campaigns </h3>
@@ -181,23 +73,8 @@ class ContributionsGraphContainer extends React.Component {
               candidates={candidates}
             />
           </div>
-        </Tab>
-      </Tabs>
-    );
-  }
-
-  render() {
-    let candidates = this.props.location.state,
-      chartData = this.state.chartData;
-
-    let candidateCampaigns = candidates.map(candidate => {
-      for (let campaign of candidate.data.campaigns) {
-        return <p> {candidate.candidateName} - {campaign.raceType} </p>;
-      }
-    });
-    // check for data integrity
-    if (chartData.length === candidates.length) {
-      return this.renderTabs(candidateCampaigns);
+        </CampaignTabs>
+      )
     } else {
       return <h1> Loading Comparison Chart </h1>;
     }
