+++ test/src/components/campaign/CampaignDetailComponentTest.js
@@ -5,10 +5,13 @@ import { Tab, Tabs } from 'react-bootstrap';
 import CampaignDetailComponent from '../../../../src/components/campaign/campaignDetailComponent';
 
 
+// mock when react router 4 is in
 // describe('<CampaignDetailComponent />', () => {
-//   it('should render 2 tabs', () => {
-//     const wrapper = shallow(<CampaignDetailComponent/>);
-//     expect(wrapper.find(Tabs)).to.have.length(1);
+//   it('should initially be loading', () => {
+//     const wrapper = shallow(<CampaignDetailComponent />);
+//     // expect(wrapper.find('<h3>')).to.have.length(1);
+//     expect(wrapper.contains(<h3>Loading...</h3>)).to.equal(true);
+//
 //   });
 //
 // });
